Refactor Translation component into separate file
